services/ingestion-service/app/core/security.py
from typing import Any

import jwt

from app.core.config import settings


def decode_access_token(token: str) -> dict[str, Any]:
    return jwt.decode(
        token,
        settings.jwt_secret,
        algorithms=[settings.jwt_algorithm],
    )
