services/ingestion-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic-settings>=2.6.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
PyJWT>=2.9.0
python-multipart>=0.0.9
redis[hiredis]>=5.0.0
